Update Section 3 - Partner Type Descriptions
Update to Partner Type Descriptions to clarify the type roles as they are applied to partner organizations. 
Partnership Committee provided the suggested updates to the Governance Committee in November 2019, and the Governance Committee reviewed the suggested updates in December 2019. 
The changes were presented and approved by the Board on January 6, 2020, at the Board Meeting held at the Winter 2020 ESIP Meeting.